Implement design and content of the second part of the MSS step
This commit implements the design and content of the second part of the
MSS step in the welcome wizard. This is a new part of the final step.
The functionality is not yet implemented and will be in a subsequent
commit.

The contents of the first part that already existed were moved to its
own file (first_part.tsx) as part of this commit as well.

[MAILPOET-4818]
