+++ mailpoet/views/mss_pitch_translations.html
@@ -1,11 +1,11 @@
 <%= localize({
-  'welcomeWizardMSSFreeTitle': _x('Connect your MailPoet account', 'Promotion for our email sending service: Title'),
-  'welcomeWizardMSSFreeSubtitle': _x('To start sending emails, create and connect a MailPoet account with your site. With a MailPoet account, you’ll get:', 'Promotion for our email sending service: Paragraph'),
+  'welcomeWizardMSSFirstPartTitle': _x('Connect your MailPoet account', 'Promotion for our email sending service: Title'),
+  'welcomeWizardMSSFirstPartSubtitle': _x('To start sending emails, create and connect a MailPoet account with your site. With a MailPoet account, you’ll get:', 'Promotion for our email sending service: Paragraph'),
   'welcomeWizardMSSList1': _x('Emails that reach inboxes, not spam boxes', 'Promotion for our email sending service: Feature item'),
   'welcomeWizardMSSList2': _x('Less hassle setting up, as we would manage all the configurations', 'Promotion for our email sending service: Feature item'),
   'welcomeWizardMSSList3Free': _x('Start sending for free and scale as you grow', 'Promotion for our email sending service: Feature item'),
   'welcomeWizardMSSList3Paid': _x('Super fast: send up to 50,000 emails per hour', 'Promotion for our email sending service: Feature item'),
-  'welcomeWizardMSSFreeButton': _x('Connect MailPoet', 'Promotion for our email sending service: Button'),
+  'welcomeWizardMSSFirstPartButton': _x('Connect MailPoet', 'Promotion for our email sending service: Button'),
   'welcomeWizardMSSAdvancedUsers': _x('(For advanced users) You can choose to use MailPoet with your own email delivery service. [link]I’ll set up my own email service[/link].', 'Promotion for our email sending service: message for users about not using MSS'),
   'welcomeWizardMSSConfirmationModalTitle': _x('Confirm sending service selection', 'Promotion for our email sending service: title of the modal where the user confirms that they want to use their own sending service'),
   'welcomeWizardMSSConfirmationModalFirstParagraph': _x('Are you sure that you would like to continue using your own email delivery service? This would require configuring your web host or a third-party email delivery service to work with the MailPoet plugin.', 'Promotion for our email sending service: first paragraph of the modal where the user confirms that they want to use their own sending service'),
@@ -13,4 +13,10 @@
   'welcomeWizardMSSConfirmationModalSecondParagraph': _x('We do not recommend this option if you are unfamiliar with setting up your own email service.', 'Promotion for our email sending service: second paragraph of the modal where the user confirms that they want to use their own sending service'),
   'welcomeWizardMSSConfirmationModalGoBackButton': _x('Go back', 'Promotion for our email sending service: go back button'),
   'welcomeWizardMSSConfirmationModalOkButton': _x('Yes, I’ll use my own service', 'Promotion for our email sending service: confirm button'),
+  'welcomeWizardMSSSecondPartTitle': _x('Activate your MailPoet account', 'Promotion for our email sending service: title of the second part'),
+  'welcomeWizardMSSSecondPartEnterKey': __('Enter your activation key to validate your account.'),
+  'welcomeWizardMSSSecondPartNoAccount': __('Don’t have one yet? [link]Sign up for a MailPoet plan[/link].'),
+  'welcomeWizardMSSSecondPartInputLabel': __('Activation key'),
+  'welcomeWizardMSSSecondPartInputPlaceholder': __('Enter your key here'),
+  'welcomeWizardMSSSecondPartButton': __('Verify account'),
 }) %>
